pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dromadaire"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "rich-pixels>=3.0.1",
    "ruff>=0.12.1",
    "textual>=3.5.0",
    "textual-dev>=1.7.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-textual-snapshot>=1.0.0",
]

[project.scripts]
app = "dromadaire.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/dromadaire"]

[dependency-groups]
dev = [
    "gitingest>=0.1.5",
]
